Refactor/HealthManageCard 관련 로직 리팩토링 (#30)
* refactor: HealthManageCard 관련로직 리팩토링

* docs: readme 수정
