+++ README.md
@@ -54,5 +54,11 @@ Eslint-config-prettier: `"^8.5.0"` <br/>
 
 ### 건강 점수 분석 결과
 
+### 맞춤 건강 관리
+
+- 사용자에게 맞춤 건강관리 서비스를 제공하는 컴포넌트입니다.
+- 항복은 총 8개입니다.
+- 복잡한 JSON 데이터를 어떻게 효율적으로 가공하여 사용할지에 대한 고민이 있었습니다.
+
 
 ## 느낀 점
