+++ src/routes/MainPage/HealthManage/index.tsx
@@ -3,7 +3,7 @@ import HealthManageCard from './HealthManageCard'
 import { getHealthManageData } from 'services/health'
 
 const HealthManage = () => {
-  const { wMymaxHscoreDy, healthMangeCardData } = getHealthManageData()
+  const { wMymaxHscoreDy, healthManageCardData } = getHealthManageData()
   const MymaxHscoreDy = wMymaxHscoreDy[Math.floor(Math.random() * wMymaxHscoreDy.length)]
 
   return (
@@ -15,14 +15,8 @@ const HealthManage = () => {
         건강점수를 최대 <span>{MymaxHscoreDy}점</span>까지 올릴 수 있어요.
       </p>
       <ul>
-        {healthMangeCardData.map((healthManageCardData, idx) => {
-          return (
-            <HealthManageCard
-              key={`key_${healthManageCardData.title}`}
-              healthManageCardData={healthManageCardData}
-              idx={idx + 1}
-            />
-          )
+        {healthManageCardData.map((item, idx) => {
+          return <HealthManageCard key={`key_${item.title}`} healthManageCardData={item} idx={idx + 1} />
         })}
       </ul>
     </section>
