+++ src/services/health.ts
@@ -171,7 +171,7 @@ export const getHealthManageData = () => {
 
   const { boj, paramMap }: { boj: IHealthManageData; paramMap: IHealthManageData } = wxcResultMap
 
-  const healthMangeCardData = property.map((value) => {
+  const healthManageCardData = property.map((value) => {
     const tag: string[] = []
     healthTagList.forEach((currentValue) => {
       if (currentValue.tagId === value) tag.push(currentValue.tag1, currentValue.tag2, currentValue.tag3)
@@ -192,6 +192,6 @@ export const getHealthManageData = () => {
 
   return {
     wMymaxHscoreDy,
-    healthMangeCardData,
+    healthManageCardData,
   }
 }
