+++ src/types/health.d.ts
@@ -2,3 +2,10 @@ export interface ScoreType {
   value?: number
   x?: string
 }
+
+export interface IHealthManageCardData {
+  title: string
+  value: string
+  boj: string[]
+  tag: string[]
+}
